server-deploy/requirements.txt
fastmcp
tavily-python
fastapi
uvicorn[standard]
pydantic
httpx
